third_party/WebKit/LayoutTests/http/tests/storage/async_cookies/document_has.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookes: cookieStore.has()</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');
  const result_after_set = await cookieStore.has('cookie-name');
  assert_equals(result_after_set, true);
}, 'cookieStore.has returns true for a cookie set by cookieStore.set()');

promise_test(async testCase => {
  await cookieStore.delete('cookie-name');
  const result_after_delete = await cookieStore.has('cookie-name');
  assert_equals(result_after_delete, false);
}, 'cookieStore.has returns false for a cookie deletedby cookieStore.delete()');

</script>
